Pdf_chatbbot/requirements.txt
streamlit==1.31.0
langchain==0.1.10
langchain-community==0.0.24
langchain-huggingface==0.0.1
langchain-chroma==0.1.0
langchain-google-genai==1.0.1
langgraph==0.0.26
chromadb==0.4.22
pypdf==4.0.1
python-dotenv==1.0.1
sentence-transformers==2.3.1
